added more fractals and UI polish

+++ app/src/main/java/val/mx/vbread/VComplex.java
@@ -32,7 +32,7 @@ public class VComplex {
                 imag * complex.real + imag * complex.real);
     }
 
-    public VComplex multiply(int fac) {
+    public VComplex multiply(double fac) {
         return new VComplex(real * fac, imag * fac);
     }
